Bug fixes, other small tweeks
Fixed precision loss of alpha values in the code-model that occured when
converting between fraction and integer representations. Alpha is now
stored as an int in 0-255 range.

Tweeked audio id buttons so that they play the sample sound at the
volume indicated by the volume slider. This allows them to be selected
without 'playing' the sound when the volume is 0.

Fixed a bug where multiple rules, groups, or filters in the treeview
could be selected at the same time after duplication.

Tweeked volume input step size to be 5 instead of 1.

Removed some old code comments.