Rule model can store multiple comments.
The rule model can now store multiple comments.

All rule lines are output in a specific order, including comments. This
is intentional, because it creates a standardized syntax for rules,
which makes them more readable. In the future I may have comments
preserve their ordering, but for now they will be inserted between the
rule name declaration and the visibility token (Show/Hide).