Fixed bug in Preview Pane + code optimizations
Fixed the example item floating outside the preview pane when it was
resized to be smaller than the item's position. The example item is now
pushed upwards when shrinking the preview area.

Moved the testing of the 'changeInProgress' flag to within the functions
that use it. This means that these functions will be called
unnecessarily but will not execute their contents. This is not ideal
from an efficiency standpoint, however, it modularizes the lock mechanic
to stay within it's context (invisible to calling code). This means that
you can call the functions from anywhere without having to remember to
test the lock first, thus making the code more maintainable. (Down with
spaghetti!)