Finished the linker, and tweeked the layout
Finished the linker, which is the js code that handles event propagation
between the treeview, code pane, input controls, rule model, and preview
pane. This is a bit complex, so I expect to find some bugs and edge
cases.

Removed remnants of React framework.

Rearranged the constraint controls and the code view line order to match
eachother.

A few other tiny code optimizations.