Added import feature and bug fixes
Added a bunch of code to allow importing filter files. This includes
some modal code which will be used elsewhere. Currently only one file
can be imported at a time, I plan to change this eventually. I also
included a sample filter in the src folder to test with.

Fixed a bug that prevented the positional sound checkbox from updating
properly when a rule was duplicated.

Made the export button show as disabled when no filter exists.